centralize the logic for pivoting to multidimensional array to represent lines/tokens
