+++ src/ast/array.js
@@ -1,3 +1,4 @@
+const getLines = require('./_get-lines')
 const SpaceError = require('../errors/parse-array-illegal-space')
 
 /**
@@ -8,7 +9,7 @@ const SpaceError = require('../errors/parse-array-illegal-space')
  */
 module.exports = function array (lines) {
 
-  let copy = lines.slice(0)
+  let copy = getLines(lines.slice(0))
   let end = copy[0].length + 1
   let value = { type: 'array', line: copy[0][0].line, column: copy[0][0].column, values: copy[0] }
